Let services use separate containers (#4)
* Let services use separate containers

Each service will have its own container so that documents will not be
mixed up. A container is named after the service it was provided for.

STORAGE_CONTAINER_NAME variable is removed as names are read from
storage account.

* Inject CloudBlobClient instead of creating

* Set storage kind to blob

No need for queues, tables, or files.

* Enable https traffic only

* Add explicit dependency between storage and container

Terraform may refresh state for a container and then re-create the
storage, removing all configured containers.