Use product variable in storage account name
Builds are failing due to error returned from Azure with code
"StorageAccountAlreadyExists". After deleting the resources the build
can be re-ran.

Each build implicitly is given information about pull request number
that is included in "product" variable. It follows "pr-{number}-rpe"
scheme.

Solution to the error is to include "product" value in the storage
account name. To make it more challenging, Azure has certain limitations
for account naming:

 - length between 3 and 24 characters
 - alphanumerics only

Because of that, any dashes in "product" value have to be removed, and
the final value has to be trimmed down to 24 characters.